swecgo: introduce Call function
By adding the Call function the client of the package is allowed to
call an arbitrary block of code within a single 'call' to the Swiss
Ephemeris C library. The results in a better predictability of when and
where the calls to the Swiss Ephemeris are made.

Call guarantees the client that it can depend on the fact if it calls
swe_calc twice those function calls are executed right after each other
without interruption of other calls to the library. This is important
because you can do some optimizations based on this. See the
Programmer's documentation paragraph 2.5 for an example.